script+cleanup(v2)
