+++ requirements.txt
@@ -48,6 +48,7 @@ google-auth-httplib2==0.2.0
 googleapis-common-protos==1.70.0
 google-cloud-core==2.4.3
 google-cloud-storage==2.14.0
+google-cloud-wallet-objects
 google-cloud-aiplatform==1.58.0
 google-cloud-firestore==2.21.0
 google-cloud-bigquery==3.35.0
@@ -60,6 +61,7 @@ grpcio-status==1.62.3
 proto-plus==1.26.1
 protobuf==4.25.8
 google-genai==1.3.0
+google-auth-oauthlib
 
 # ============================ Firebase Admin SDK ==============================
 firebase-admin==6.4.0
