+++ utils/utils.py
@@ -1,11 +1,33 @@
-from google.cloud import firestore, storage
 import os
-import json
+from firebase_admin import credentials
+import firebase_admin
+from google.cloud import firestore, storage
 import tempfile
 import logging
 import atexit
+import json
+import re
 
 logger = logging.getLogger(__name__)
+
+def get_credentials():
+    """Load credentials from environment variables."""
+
+    try:
+        PROJECT_ID = os.getenv("PROJECT_ID")
+        LOCATION = os.getenv("GCP_LOCATION")
+        GOOGLE_WALLET_ISSUER_ID = os.getenv("GOOGLE_WALLET_ISSUER_ID")
+    except KeyError:
+        raise RuntimeError("GCP_PROJECT_ID not found in .env file.")
+
+    #Initialize cloud storage and firestore clients
+
+    try:
+        BUCKET_NAME = os.getenv("GCS_BUCKET_NAME")
+    except KeyError:
+        raise RuntimeError("GCS_BUCKET_NAME not found in .env file. Please set it.")
+
+    return PROJECT_ID, LOCATION, BUCKET_NAME, GOOGLE_WALLET_ISSUER_ID
 
 # Global variable to keep track of temporary credentials file
 _temp_credentials_path = None
@@ -97,3 +119,32 @@ def initialize_gcs_client():
     except Exception as e:
         logger.error(f"GCS client initialization error: {e}")
         raise
+
+def parse_json(raw_text):
+    """
+    Extracts and parses a JSON object from a Markdown code block (e.g., ```json ... ```).
+
+    Args:
+        raw_text (str): Raw text that includes a Markdown-style JSON block.
+
+    Returns:
+        dict: Parsed JSON data as a Python dictionary.
+
+    Raises:
+        ValueError: If no valid JSON block is found or JSON parsing fails.
+    """
+    if not isinstance(raw_text, str):
+        raise ValueError("Input must be a string")
+
+    # Match ```json ... ```
+    match = re.search(r"```(?:json)?\s*(\{.*?\})\s*```", raw_text, re.DOTALL)
+    if not match:
+        raise ValueError("No valid JSON block found in the input text")
+
+    json_text = match.group(1).strip()
+
+    try:
+        return json.loads(json_text)
+    except json.JSONDecodeError as e:
+        raise ValueError(f"Failed to parse JSON: {e}")
+
